Implement experimental support for S130 on nRF51
This is completely untested.  The S130 API is a mix of nRF51-S110 and
nRF52-S132 calls.  So where necessary, the old checks for the nRF51
remain.

S110 is left as part of the codebase because S130 supports only revision
3 nRF51 devices.  Other than inspecting the particular SoC on your
board, there does not appear to be a good way to tell which boards have
which device, as most manufacturers make no guarantee.